Broken commit: Move folding-related scripting API to another module
This commit only copies the relevant lines from PoseEditMode.ts to ScriptsApi.ts - it doesn't modify them, only adds code around them.

Because of that, ScriptsApi.ts is broken for this commit.
